Create bark.ts
The Solana Token 2022 "TransferFee" extension allows you to configure a transfer fee directly on the Mint Account, enabling fees to be collected at a protocol level. Every time tokens are transferred, the fee is set aside in the recipient's Bark´s Token Account. This fee is untouchable by the recipient and can only be accessed by the withdrawal authority.

The design of pooling transfer fees at the recipient account is meant to maximize the parallelization of transactions. Otherwise, one configured fee recipient account would be write-locked between parallel transfers, decreasing the throughput of the Bark protocol.